workflow-backend/app/workers/smart_tiling.py
# ============================================
# FILE: workflow-backend/app/workers/smart_tiling.py
# ============================================
"""
Smart Tiling with Tissue Detection
Key Innovation: Skip 60-70% of background tiles
"""

# Lazy import openslide to avoid errors if system library not installed
try:
    import openslide
    OPENSLIDE_AVAILABLE = True
except ImportError:
    OPENSLIDE_AVAILABLE = False
    openslide = None

import numpy as np
from skimage import filters
import cv2
from typing import List, Tuple, Dict
import logging

logger = logging.getLogger(__name__)


class SmartTiler:
    """
    Intelligent tiling strategy for whole-slide images
    
    Innovations:
    1. Tissue detection at thumbnail level (fast)
    2. Skip background tiles (major speedup)
    3. Overlap with Gaussian blending (no seams)
    """
    
    def __init__(self, tile_size: int = 1024, overlap: int = 128, min_tissue_ratio: float = 0.1):
        """
        Args:
            tile_size: Size of each tile (pixels)
            overlap: Overlap between adjacent tiles (pixels)
            min_tissue_ratio: Minimum tissue ratio to include tile (0-1)
        """
        self.tile_size = tile_size
        self.overlap = overlap
        self.min_tissue_ratio = min_tissue_ratio
        
        logger.info(f"SmartTiler initialized: tile_size={tile_size}, overlap={overlap}")
    
    
    def detect_tissue_fast(self, wsi_path: str) -> Tuple[np.ndarray, Tuple[int, int]]:
        """
        Fast tissue detection using thumbnail
        
        Strategy:
        1. Load low-resolution thumbnail (much faster)
        2. Convert to grayscale
        3. Apply Otsu thresholding (tissue is darker than background)
        
        Returns:
            (tissue_mask, full_dimensions)
            tissue_mask: Binary mask at thumbnail resolution
            full_dimensions: (width, height) at full resolution
        """
        logger.info(f"Loading WSI: {wsi_path}")
        slide = openslide.OpenSlide(wsi_path)
        
        # Get thumbnail (e.g., 2000x2000 instead of 50000x50000)
        thumbnail = slide.get_thumbnail((2000, 2000))
        logger.info(f"Thumbnail size: {thumbnail.size}, Full size: {slide.dimensions}")
        
        # Convert to grayscale
        gray = cv2.cvtColor(np.array(thumbnail), cv2.COLOR_RGB2GRAY)
        
        # Otsu thresholding - tissue is typically darker
        try:
            threshold = filters.threshold_otsu(gray)
            tissue_mask = gray < threshold  # Tissue pixels are below threshold
        except Exception as e:
            logger.warning(f"Otsu failed, using manual threshold: {e}")
            # Fallback: assume tissue is darker than 200
            tissue_mask = gray < 200
        
        tissue_percentage = tissue_mask.mean() * 100
        logger.info(f"Tissue coverage: {tissue_percentage:.1f}%")
        
        return tissue_mask, slide.dimensions
    
    
    def generate_smart_tiles(self, wsi_path: str) -> List[Dict]:
        """
        Generate tiles ONLY where tissue exists
        
        Returns:
            List of tile info dicts: [{x, y, width, height}, ...]
        """
        tissue_mask, (full_width, full_height) = self.detect_tissue_fast(wsi_path)
        
        # Calculate scale factors between thumbnail and full resolution
        scale_x = tissue_mask.shape[1] / full_width
        scale_y = tissue_mask.shape[0] / full_height
        
        logger.debug(f"Scale factors: x={scale_x:.6f}, y={scale_y:.6f}")
        
        # Generate tiles
        tiles = []
        stride = self.tile_size - self.overlap
        
        total_possible = 0
        skipped = 0
        
        for y in range(0, full_height, stride):
            for x in range(0, full_width, stride):
                total_possible += 1
                
                # Map to thumbnail coordinates
                thumb_x = int(x * scale_x)
                thumb_y = int(y * scale_y)
                thumb_w = int(self.tile_size * scale_x)
                thumb_h = int(self.tile_size * scale_y)
                
                # Ensure we don't go out of bounds
                thumb_x = min(thumb_x, tissue_mask.shape[1] - 1)
                thumb_y = min(thumb_y, tissue_mask.shape[0] - 1)
                thumb_w = min(thumb_w, tissue_mask.shape[1] - thumb_x)
                thumb_h = min(thumb_h, tissue_mask.shape[0] - thumb_y)
                
                # Sample tissue mask in this region
                tile_region = tissue_mask[
                    thumb_y:thumb_y+thumb_h,
                    thumb_x:thumb_x+thumb_w
                ]
                
                # Check if sufficient tissue present
                tissue_ratio = tile_region.mean() if tile_region.size > 0 else 0
                
                if tissue_ratio >= self.min_tissue_ratio:
                    # Include this tile
                    tiles.append({
                        'x': x,
                        'y': y,
                        'width': min(self.tile_size, full_width - x),
                        'height': min(self.tile_size, full_height - y),
                        'tissue_ratio': float(tissue_ratio)
                    })
                else:
                    skipped += 1
        
        skip_percentage = (skipped / total_possible) * 100 if total_possible > 0 else 0
        
        logger.info(f"Tile generation complete:")
        logger.info(f"  Total possible tiles: {total_possible}")
        logger.info(f"  Tiles with tissue: {len(tiles)}")
        logger.info(f"  Tiles skipped: {skipped} ({skip_percentage:.1f}%)")
        logger.info(f"  SPEEDUP: Processing only {len(tiles)}/{total_possible} tiles!")
        
        return tiles
    
    
    def create_blend_weights(self, tile_size: int, overlap: int) -> np.ndarray:
        """
        Create Gaussian blending weights for tile overlap regions
        
        Prevents visible seams at tile boundaries
        
        Returns:
            2D array of weights (1.0 in center, fades to 0 at edges)
        """
        weights = np.ones((tile_size, tile_size), dtype=np.float32)
        
        if overlap == 0:
            return weights
        
        # Create fade in overlap regions
        for i in range(overlap):
            fade = i / overlap  # 0 to 1
            
            # Top edge
            weights[i, :] *= fade
            # Bottom edge
            weights[-(i+1), :] *= fade
            # Left edge
            weights[:, i] *= fade
            # Right edge
            weights[:, -(i+1)] *= fade
        
        return weights
    
    
    def merge_tiles_with_blending(
        self, 
        tile_results: List[Dict], 
        full_size: Tuple[int, int]
    ) -> np.ndarray:
        """
        Merge overlapping tiles with Gaussian blending
        
        Args:
            tile_results: List of dicts with 'x', 'y', 'mask' keys
            full_size: (width, height) of output
        
        Returns:
            Merged mask as numpy array
        """
        logger.info(f"Merging {len(tile_results)} tiles into {full_size}")

        # Initialize accumulation arrays
        # full_size is (width, height) from OpenSlide, but numpy needs (height, width)
        result = np.zeros((full_size[1], full_size[0]), dtype=np.float32)
        weights_sum = np.zeros((full_size[1], full_size[0]), dtype=np.float32)
        
        # Create blending weights once
        blend_weights = self.create_blend_weights(self.tile_size, self.overlap)
        
        # Merge each tile
        for i, tile_result in enumerate(tile_results):
            x = tile_result['x']
            y = tile_result['y']
            mask = tile_result['mask']
            
            h, w = mask.shape
            
            # Get appropriate weights for this tile size
            if h != self.tile_size or w != self.tile_size:
                # Tile is at edge, resize weights
                tile_weights = cv2.resize(
                    blend_weights, 
                    (w, h), 
                    interpolation=cv2.INTER_LINEAR
                )
            else:
                tile_weights = blend_weights[:h, :w]
            
            # Accumulate with blending
            result[y:y+h, x:x+w] += mask * tile_weights
            weights_sum[y:y+h, x:x+w] += tile_weights
            
            if (i + 1) % 100 == 0:
                logger.debug(f"Merged {i+1}/{len(tile_results)} tiles")
        
        # Normalize by weights (avoid division by zero)
        result = np.divide(
            result, 
            weights_sum,
            out=np.zeros_like(result),
            where=weights_sum > 1e-8
        )
        
        logger.info("Tile merging complete")
        
        return result
    
    
    def save_tissue_mask_visualization(self, wsi_path: str, output_path: str):
        """
        Save visualization of tissue detection (for debugging/demo)
        """
        tissue_mask, dimensions = self.detect_tissue_fast(wsi_path)
        
        # Convert mask to RGB
        vis = np.zeros((*tissue_mask.shape, 3), dtype=np.uint8)
        vis[tissue_mask] = [0, 255, 0]  # Green for tissue
        vis[~tissue_mask] = [255, 0, 0]  # Red for background
        
        cv2.imwrite(output_path, vis)
        logger.info(f"Saved tissue mask visualization to {output_path}")
