workflow-backend/requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
ImageIO==2.37.2
instanseg>=0.1.0
kombu==5.5.4
lazy_loader==0.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
openslide-python==1.4.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==7.0.1
requests>=2.31.0
scikit-image==0.25.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tifffile==2025.10.16
torch>=2.0.0
torchvision>=0.15.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
